refactor(fonts): unify the duplicated glyph decoder
The glyph decoder existed three times. Extraction and rendering each
carried their own copy of the byte-mode choice, the character iterator
and the Unicode fallback table, and the copies had drifted: the render
path's fallback table resolved a fifth of the codes the extraction one
did, and its byte-mode choice ignored a ToUnicode CMap that declares a
two-byte codespace, so a CJK font whose /Encoding name says otherwise
decoded one byte at a time while extraction read it correctly.

One decoder now lives in fonts::unicode_decode and both paths call it.
What differs between them is stated as policy rather than reimplemented:
extraction keeps an unmapped code and writes a question mark for one no
table resolves, rendering expands a ligature for the shaper and paints
nothing for a code it cannot resolve.

The render path keeps its own drop tally. That decode runs before the
run is routed, and the byte-to-GID and direct-CID paths paint correctly
from raw bytes for exactly the font classes whose Unicode decode fails,
so only the chosen paint path records a drop.
