+++ src/rendering/text_rasterizer.rs
@@ -15,6 +15,9 @@ use crate::content::operators::TextElement;
 use crate::content::GraphicsState;
 use crate::document::PdfDocument;
 use crate::error::{Error, Result};
+use crate::fonts::unicode_decode::{
+    char_codes_with_widths, get_byte_mode, ByteMode, DecodePolicy, TextCharIter,
+};
 use crate::object::Object;
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -712,75 +715,30 @@ impl TextRasterizer {
     }
 
     /// Decode raw PDF text bytes to a Unicode string based on font type.
+    /// Decode a show string the way the render path paints it.
+    ///
+    /// Rendering expands a ligature so the shaper sees ordinary glyphs, drops
+    /// a code no table resolves rather than writing a question mark into the
+    /// page, and never keeps an unmapped code.
+    ///
+    /// No GlyphDropTally here: this decode runs before render_text routes the
+    /// run, and the byte-to-GID / direct-CID paths paint correctly from raw
+    /// bytes for exactly the font classes whose Unicode decode fails. Only the
+    /// chosen paint path may record a drop.
     fn decode_text_to_unicode(
         &self,
         bytes: &[u8],
         font: Option<&crate::fonts::FontInfo>,
     ) -> String {
-        // No GlyphDropTally here: this decode runs before render_text routes
-        // the run, and the byte-to-GID / direct-CID paths paint correctly from
-        // raw bytes for exactly the font classes whose Unicode decode fails.
-        // Only the chosen paint path may record a drop.
-        let raw_result = if let Some(font) = font {
-            let mut result = String::new();
-            // Use pre-computed lookup table for performance if it's a simple font
-            if font.subtype != "Type0" {
-                let table = font.get_byte_to_char_table();
-                for &byte in bytes {
-                    let c = table[byte as usize];
-                    if c != '\0' {
-                        result.push(c);
-                    } else {
-                        // Fallback: multi-char mapping or unmapped byte
-                        let char_str = font
-                            .char_to_unicode(byte as u32)
-                            .unwrap_or_else(|| fallback_char_to_unicode(byte as u32));
-                        if char_str != "\u{FFFD}" {
-                            result.push_str(&char_str);
-                        }
-                    }
-                }
-            } else {
-                // Complex font: use unified iterator for robust multi-byte decoding
-                for (char_code, _) in TextCharIter::new(bytes, Some(font)) {
-                    let char_str = font
-                        .char_to_unicode(char_code as u32)
-                        .unwrap_or_else(|| fallback_char_to_unicode(char_code as u32));
-
-                    if char_str != "\u{FFFD}" {
-                        result.push_str(&char_str);
-                    }
-                }
-            }
-            result
-        } else {
-            // No font - fallback to Latin-1 (ISO 8859-1) encoding
-            bytes.iter().map(|&b| char::from(b)).collect()
-        };
-
-        // Filter control characters from failed encoding resolution,
-        // and expand presentation-form ligature code points (fi, fl, ffi,
-        // ffl, st, ct, …) into their component letters so the shaper
-        // passes the cluster through as ordinary glyphs instead of
-        // dropping it or producing a lone box. `extract_text` already
-        // does this on the extraction path via
-        // `ligature_processor::get_ligature_components`; without the
-        // same decomposition on the render path, words like
-        // "Efficient" rasterize as "Effi  ert" because the shaper can't
-        // resolve the ligature cluster against the fallback system
-        // font. See issue #331 (R2).
-        let mut filtered = String::with_capacity(raw_result.len());
-        for c in raw_result.chars() {
-            if c < '\x20' && c != '\t' && c != '\n' && c != '\r' {
-                continue;
-            }
-            if let Some(components) = crate::text::ligature_processor::get_ligature_components(c) {
-                filtered.push_str(components);
-            } else {
-                filtered.push(c);
-            }
-        }
-        filtered
+        crate::fonts::unicode_decode::decode_text_to_unicode(
+            bytes,
+            font,
+            DecodePolicy {
+                preserve_unmapped: false,
+                decompose_ligatures: true,
+                question_mark_for_invalid: false,
+            },
+        )
     }
 
     /// Measure-only: compute the horizontal advance of a Tj text string
@@ -1975,136 +1933,6 @@ impl TextRasterizer {
     }
 }
 
-/// Byte grouping mode for CID font character code decoding.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ByteMode {
-    /// Single-byte codes (simple fonts, some predefined CMaps)
-    OneByte,
-    /// Always 2-byte codes (Identity-H/V, UCS2)
-    TwoByte,
-    /// Shift-JIS variable-width (1 or 2 bytes depending on lead byte)
-    ShiftJIS,
-}
-
-/// Get byte grouping mode for a font.
-fn get_byte_mode(font: Option<&crate::fonts::FontInfo>) -> ByteMode {
-    if let Some(font) = font {
-        if font.subtype == "Type0" {
-            match &font.encoding {
-                crate::fonts::Encoding::Identity => ByteMode::TwoByte,
-                crate::fonts::Encoding::Standard(name) => {
-                    if (name.contains("Identity") && !name.contains("OneByteIdentity"))
-                        || name.contains("UCS2")
-                        || name.contains("UTF16")
-                    {
-                        ByteMode::TwoByte
-                    } else if name.contains("RKSJ") {
-                        ByteMode::ShiftJIS
-                    } else if name.contains("EUC")
-                        || name.contains("GBK")
-                        || name.contains("GBpc")
-                        || name.contains("GB-")
-                        || name.contains("CNS")
-                        || name.contains("B5")
-                        || name.contains("KSC")
-                        || name.contains("KSCms")
-                    {
-                        ByteMode::TwoByte
-                    } else {
-                        ByteMode::OneByte
-                    }
-                },
-                _ => ByteMode::OneByte,
-            }
-        } else {
-            ByteMode::OneByte
-        }
-    } else {
-        ByteMode::OneByte
-    }
-}
-
-/// Iterator over characters in a PDF string based on font encoding.
-struct TextCharIter<'a> {
-    bytes: &'a [u8],
-    byte_mode: ByteMode,
-    index: usize,
-}
-
-impl<'a> TextCharIter<'a> {
-    fn new(bytes: &'a [u8], font: Option<&crate::fonts::FontInfo>) -> Self {
-        Self {
-            bytes,
-            byte_mode: get_byte_mode(font),
-            index: 0,
-        }
-    }
-}
-
-impl<'a> Iterator for TextCharIter<'a> {
-    type Item = (u16, usize); // (char_code, bytes_consumed)
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.index >= self.bytes.len() {
-            return None;
-        }
-
-        let (char_code, bytes_consumed) = match self.byte_mode {
-            ByteMode::TwoByte if self.index + 1 < self.bytes.len() => {
-                (((self.bytes[self.index] as u16) << 8) | (self.bytes[self.index + 1] as u16), 2)
-            },
-            ByteMode::ShiftJIS => {
-                let b = self.bytes[self.index];
-                let is_lead = (0x81..=0x9F).contains(&b) || (0xE0..=0xFC).contains(&b);
-                if is_lead && self.index + 1 < self.bytes.len() {
-                    (((b as u16) << 8) | (self.bytes[self.index + 1] as u16), 2)
-                } else {
-                    (b as u16, 1)
-                }
-            },
-            _ => (self.bytes[self.index] as u16, 1),
-        };
-
-        self.index += bytes_consumed;
-        Some((char_code, bytes_consumed))
-    }
-}
-
-/// Fallback function to map common character codes to Unicode when ToUnicode CMap fails.
-fn fallback_char_to_unicode(char_code: u32) -> String {
-    match char_code {
-        0x2014 => "—".to_string(),
-        0x2013 => "–".to_string(),
-        0x2018 => "\u{2018}".to_string(),
-        0x2019 => "\u{2019}".to_string(),
-        0x201C => "\u{201C}".to_string(),
-        0x201D => "\u{201D}".to_string(),
-        0x2022 => "•".to_string(),
-        0x2026 => "…".to_string(),
-        0x00B0 => "°".to_string(),
-        0x00B1 => "±".to_string(),
-        0x00D7 => "×".to_string(),
-        0x00F7 => "÷".to_string(),
-        0x2202 => "∂".to_string(),
-        0x2207 => "∇".to_string(),
-        0x220F => "∏".to_string(),
-        0x2211 => "∑".to_string(),
-        0x221A => "√".to_string(),
-        0x221E => "∞".to_string(),
-        0x2260 => "≠".to_string(),
-        0x2261 => "≡".to_string(),
-        0x2264 => "≤".to_string(),
-        0x2265 => "≥".to_string(),
-        code => {
-            if let Some(ch) = char::from_u32(code) {
-                ch.to_string()
-            } else {
-                "\u{FFFD}".to_string()
-            }
-        },
-    }
-}
-
 impl Default for TextRasterizer {
     fn default() -> Self {
         Self::new()
@@ -2141,7 +1969,12 @@ fn measure_text_bytes(
     let mut advance: f32 = 0.0;
 
     if let Some(font) = font_info {
-        for (char_code, nbytes) in TextCharIter::new(bytes, Some(font)) {
+        // `char_codes_with_widths` keeps the segmentation identical to the
+        // painted path's width lookups (UTF-8 CMaps are variable-width), so
+        // the measured advance matches what rendering would have produced,
+        // while still exposing each code's byte width for the Tw gate below.
+        for (code, nbytes) in char_codes_with_widths(bytes, font) {
+            let char_code = u16::try_from(code).unwrap_or(0);
             // Per ISO 32000-1 §9.4.4 the advance formula differs by writing
             // mode:
             //   horizontal: tx = ((w0 * Tfs) + Tc + Tw) * Th
@@ -2151,8 +1984,10 @@ fn measure_text_bytes(
             // vertical Tc / Tw.
             // Per §9.3.3, Tw applies only to the single-byte code 32 — a
             // 2-byte CID 0x0020 under Identity-H/V or another multi-byte
-            // CMap must not take Tw.
-            let word_space_eligible = nbytes == 1 && char_code == 0x20;
+            // CMap must not take Tw. This also covers UTF-8-codespace CMaps:
+            // a UTF-8 code 0x20 with width 1 is exactly the plain-ASCII-space
+            // case Tw is meant to cover.
+            let word_space_eligible = nbytes == 1 && code == 0x20;
             if wmode == 0 {
                 let glyph_adv = font.get_glyph_width(char_code) * font_size / 1000.0;
                 advance += (glyph_adv + gs.char_space) * h_scale;
@@ -2567,6 +2402,48 @@ mod tests {
         assert!(
             (half - 10.0).abs() < 0.01,
             "Th=50% must halve the returned advance (§9.4.4 tx·Th): got {half}, full was {full}"
+        );
+    }
+
+    /// A Type0 font whose CMap uses the UTF-8-codespace convention, as
+    /// opposed to `make_vertical_test_font`'s fixed 2-byte Identity CMap.
+    fn make_utf8_cmap_test_font() -> FontInfo {
+        let mut font = make_vertical_test_font();
+        font.encoding = Encoding::Standard("UniFull-UTF8-H".to_string());
+        font.wmode = 0;
+        font
+    }
+
+    /// `measure_text_bytes` must segment with `char_codes_with_widths`, so a
+    /// UTF-8-codespace CMap's single-byte code 0x20 still receives Tw
+    /// (ISO 32000-1 §9.3.3) even though the font routes through the UTF-8
+    /// branch rather than `TextCharIter`.
+    #[test]
+    fn measure_text_bytes_applies_tw_for_utf8_cmap_single_byte_space() {
+        let font = make_utf8_cmap_test_font();
+        let mut gs = GraphicsState::new();
+        gs.font_size = 12.0;
+        gs.text_wmode = 0;
+        gs.word_space = 100.0;
+
+        // 0xC3 0xA9 ("é", 2-byte UTF-8 code) followed by a literal 1-byte
+        // space (0x20) — the space must stay segmented as width 1 and
+        // receive Tw despite following a multi-byte code in the same run.
+        let bytes: &[u8] = &[0xC3, 0xA9, 0x20];
+        let advance = measure_text_bytes(bytes, &gs, Some(&font));
+
+        // Glyph widths only matter as a constant offset here; what this
+        // test asserts is the *delta* Tw contributes — isolate it by
+        // diffing against the same bytes with word_space forced to 0.
+        let mut gs_no_tw = gs.clone();
+        gs_no_tw.word_space = 0.0;
+        let advance_no_tw = measure_text_bytes(bytes, &gs_no_tw, Some(&font));
+
+        assert!(
+            (advance - advance_no_tw - 100.0).abs() < 0.01,
+            "Tw must apply once for the trailing single-byte space in a UTF-8-CMap run, \
+             got delta {}",
+            advance - advance_no_tw
         );
     }
 }
